requirements.txt
pandas~=1.4.1
numpy~=1.22.3
scikit-learn~=1.0.2
nltk~=3.7
dask~=2022.3.0
pyarrow~=7.0.0
kaggle~=1.5.12
click~=8.0.4
spacy~=3.2.3
fastapi~=0.75.0
uvicorn~=0.17.6
lxml~=4.8.0
beautifulsoup4~=4.10.0
python-dotenv~=0.20.0
elasticsearch[async]~=8.1.1
autoflake~=1.4
autopep8~=1.6.0
mypy~=0.942
mypy-extensions~=0.4.3
black~=22.3.0
requests~=2.27.1
types-requests~=2.27.16
